test: mock api call to pass walletLogin tests
